+++ packages/orq-rc/src/models/operations/filelist.ts
@@ -157,7 +157,7 @@ export const FileListData$inboundSchema: z.ZodType<
   file_name: z.string(),
   workspace_id: z.string(),
   created: z.string().datetime({ offset: true }).default(
-    "2025-11-17T07:33:22.581Z",
+    "2025-11-17T09:11:42.760Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {
@@ -190,7 +190,7 @@ export const FileListData$outboundSchema: z.ZodType<
   bytes: z.number(),
   fileName: z.string(),
   workspaceId: z.string(),
-  created: z.date().default(() => new Date("2025-11-17T07:33:22.581Z"))
+  created: z.date().default(() => new Date("2025-11-17T09:11:42.760Z"))
     .transform(v => v.toISOString()),
 }).transform((v) => {
   return remap$(v, {