+++ packages/orq-rc/RELEASES.md
@@ -5049,3 +5049,13 @@ Based on:
 - [typescript v4.0.0-rc.38] packages/orq-rc
 ### Releases
 - [NPM v4.0.0-rc.38] https://www.npmjs.com/package/@orq-ai/node/v/4.0.0-rc.38 - packages/orq-rc
+
+## 2025-11-17 09:13:18
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.658.1 (2.755.6) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.0.0-rc.39] packages/orq-rc
+### Releases
+- [NPM v4.0.0-rc.39] https://www.npmjs.com/package/@orq-ai/node/v/4.0.0-rc.39 - packages/orq-rc